+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to the "Shell Script Snippets" extension will be documented 
 ## [0.16.0] - 2023-04-02
 
 ### Changed
-- Added `l13SHSnippets.useFunctionBlockPadding` to enable or disable empty lines in function block scopes.
+- Added `l13Snippets.shellScript.useFunctionBlockPadding` to enable or disable line padding in function blocks.
 
 ## [0.15.1] - 2021-02-01